Refactored CreateWallet scene, fixed endless loading problem, added ability to show wallet error if that happening
